Moved denicker to allow for easier adding of others
