+++ src/main/java/com/roxiun/mellow/feature/nicks/NumberDenicker.java
@@ -133,21 +133,8 @@ public class NumberDenicker {
 
         AsyncExecutor.getInstance().profileIo(() -> {
             try {
-                int[] rangeValues = { 0, 50, 100, 200, 500, 1000 };
-                int[] maxValues = { 5, 10, 20 };
-
-                int rangeIndex = type.equals("finals")
-                    ? config.finalsRange
-                    : config.bedsRange;
-                int maxIndex = config.maxResults;
-
-                if (
-                    rangeIndex < 0 || rangeIndex >= rangeValues.length
-                ) rangeIndex = 1; // Default to 200
-                if (maxIndex < 0 || maxIndex >= maxValues.length) maxIndex = 0; // Default to 5
-
-                int range = rangeValues[rangeIndex];
-                int max = maxValues[maxIndex];
+                int range = config.getAuroraDenickRange(type);
+                int max = config.getAuroraDenickMaxResults();
 
                 AuroraApi.AuroraResponse response = auroraApi.queryStats(
                     type,
